refactor base validators and claim validators
rename vaildator files and classes and combine where
appropriate in preparation for implementing subclassing
validators for agfs/lgs claim splitting.
